Move product auxiliary function to itertools::detail
